refactor: created BenecifiaryCard.jsx component for display on BeneficiaryStories.js
